combine 3 files -> test_assessments.py
* test_retrieve_update_destroy_assessment.py
* test_create_assessments.py
* test_list_assessments.py

note that I've just identified that `test_list_create_assessments.py` is
a left-over duplicate: it previously got split into
test_create_assessments and test_list_assessments. I'll delete it in a
subsequent commit.
